+++ packages/recommend-core/src/index.ts
@@ -6,5 +6,5 @@ export * from './getLookingSimilar';
 export * from './getTrendingFacets';
 export * from './getTrendingItems';
 export * from './getRecommendedForYou';
-export * from './personalization';
+export * from './experimental-personalization';
 export * from './types';
